move limited content-setting code to builder
