+++ dockstore-webservice/src/main/java/io/dockstore/webservice/helpers/SourceFileHelper.java
@@ -17,13 +17,9 @@
 
 package io.dockstore.webservice.helpers;
 
-import com.google.common.primitives.Bytes;
-import io.dockstore.common.DescriptorLanguage;
 import io.dockstore.webservice.core.SourceFile;
-import java.nio.charset.StandardCharsets;
 import java.util.Map;
 import java.util.Optional;
-import org.apache.commons.lang3.StringUtils;
 import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -33,47 +29,10 @@ import org.yaml.snakeyaml.constructor.SafeConstructor;
 
 public final class SourceFileHelper {
 
-    private static final long BYTES_PER_MEGABYTE = 1024L * 1024L;
-    private static final long MAXIMUM_FILE_SIZE = BYTES_PER_MEGABYTE;
-    private static final long NOTEBOOK_MAXIMUM_FILE_SIZE = 3 * BYTES_PER_MEGABYTE;
     private static final Logger LOGGER = LoggerFactory.getLogger(SourceFileHelper.class);
 
     private SourceFileHelper() {
 
-    }
-
-    public static void setContentWithLimits(SourceFile file, String content, String path) {
-        String limitedContent = content;
-        if (content != null) {
-            byte[] bytes = content.getBytes(StandardCharsets.UTF_8);
-            long maximumSize = computeMaximumSize(path);
-            // A large file is probably up to no good.
-            if (bytes.length > maximumSize) {
-                limitedContent = String.format("Dockstore does not store files of this type over %.1fMB in size", maximumSize / (double) BYTES_PER_MEGABYTE);
-            }
-            // Postgresql cannot store strings that contain nul characters.
-            if (Bytes.indexOf(bytes, Byte.decode("0x00")) != -1) {
-                limitedContent = "Dockstore does not store binary files";
-            }
-        }
-        file.setContent(limitedContent);
-    }
-
-    private static long computeMaximumSize(String path) {
-        // Jupyter notebooks can contain embedded images, making them larger, on average.
-        if (StringUtils.endsWith(path, ".ipynb")) {
-            return NOTEBOOK_MAXIMUM_FILE_SIZE;
-        }
-        return MAXIMUM_FILE_SIZE;
-    }
-
-    public static SourceFile create(DescriptorLanguage.FileType type, String content, String path, String absolutePath) {
-        SourceFile file = new SourceFile();
-        file.setType(type);
-        file.setPath(path);
-        file.setAbsolutePath(absolutePath);
-        setContentWithLimits(file, content, absolutePath);
-        return file;
     }
 
     public static SourceFile duplicate(SourceFile src) {
